fix(model): breakbar_damage is f64 not u64
EI emits breakbar damage as a fractional value (e.g. 15.2) computed
across time slices, not whole hits. The original u64 typing failed
mid-deserialise on real WvW logs. Confirmed by running EI manually on
20260425-195341.zevtc — fractional 15.2 appears in DpsAll.

+++ src/ei_model.rs
@@ -83,8 +83,10 @@ pub struct EiPlayer {
 pub struct DpsAll {
     pub damage: u64,
     pub dps: u64,
+    /// EI emits this as a float (e.g. `15.2`) — breakbar damage is
+    /// computed across fractional time slices, not whole hits.
     #[serde(default)]
-    pub breakbar_damage: u64,
+    pub breakbar_damage: f64,
 }
 
 #[derive(Debug, Clone, Default, Deserialize)]
